fix(warnings): Silence msvc type noise for SoundFileBridge

+++ src/MayaFlux/Buffers/Container/SoundFileBridge.cpp
@@ -21,7 +21,7 @@ SoundFileBridge::SoundFileBridge(uint32_t channel_id,
 void SoundFileBridge::setup_processors(Buffers::ProcessingToken token)
 {
     m_capture_stream = std::make_shared<Kakshya::DynamicSoundStream>(
-        s_registered_sample_rate,
+        static_cast<uint32_t>(s_registered_sample_rate),
         static_cast<uint32_t>(get_container()->get_structure().get_channel_count()));
 
     m_stream_writer = std::make_shared<SoundStreamWriter>(m_capture_stream);
